pyproject.toml
[tool.poetry]
name = "text-to-sql-rag"
version = "0.1.0"
description = "Agentic text-to-SQL RAG solution with LlamaIndex and OpenSearch"
authors = ["User <user@example.com>"]
readme = "README.md"
packages = [{include = "text_to_sql_rag", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
llama-index = "^0.12.52"
llama-index-vector-stores-opensearch = "^0.5.6"
llama-index-embeddings-bedrock = "^0.5.2"
llama-index-llms-bedrock = "^0.3.8"
langgraph = "^0.0.69"
opensearch-py = "^2.4.0"
boto3 = "^1.34.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
redis = "^5.0.0"
pyyaml = "^6.0.1"
httpx = "^0.26.0"
structlog = "^23.2.0"
pymongo = "^4.6.0"
motor = "^3.3.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true